truncate timesteps to be ints

+++ analysis/src/types/vehicle_lane_hist.cc
@@ -17,7 +17,7 @@ namespace types {
    * @param segment_id segment id
    * @param timestep   current timestep
    */
-  vehicle_lane_hist_t::vehicle_lane_hist_t(const std::string& segment_id, double timestep)
+  vehicle_lane_hist_t::vehicle_lane_hist_t(const std::string& segment_id, int timestep)
     : segments() {
     this->at_segment(segment_id, timestep);
   }
@@ -27,7 +27,7 @@ namespace types {
    * @param segment_id the segment the vehicle is currently on
    * @param timestep   the current simulation timestep
    */
-  void vehicle_lane_hist_t::at_segment(const std::string& segment_id, double timestep) {
+  void vehicle_lane_hist_t::at_segment(const std::string& segment_id, int timestep) {
     this->segments.insert(std::make_pair(segment_id, timestep));
   }
 
@@ -37,8 +37,8 @@ namespace types {
    * @param  current_ts the current timestep
    * @return            the number of timesteps since being on that segment or -1 if never seen
    */
-  double vehicle_lane_hist_t::timesteps_since_seen(const std::string& segment_id, double current_ts) const {
-    std::unordered_map<std::string,double>::const_iterator it
+  int vehicle_lane_hist_t::timesteps_since_seen(const std::string& segment_id, int current_ts) const {
+    std::unordered_map<std::string,int>::const_iterator it
       = segments.find(segment_id);
 
     if (it != segments.end()) {
