+++ analysis/src/types/vehicle_lane_hist.h
@@ -17,7 +17,7 @@ namespace types {
   struct vehicle_lane_hist_t {
   private:
     //the segments the vehicle has visited and when
-    std::unordered_map<std::string,double> segments;
+    std::unordered_map<std::string,int> segments;
 
   public:
     /**
@@ -30,7 +30,7 @@ namespace types {
      * @param segment_id segment id
      * @param timestep   current timestep
      */
-    vehicle_lane_hist_t(const std::string& segment_id, double timestep);
+    vehicle_lane_hist_t(const std::string& segment_id, int timestep);
 
     //no copy
     vehicle_lane_hist_t(const vehicle_lane_hist_t&) = delete;
@@ -41,7 +41,7 @@ namespace types {
      * @param segment_id the segment the vehicle is currently on
      * @param timestep   the current simulation timestep
      */
-    void at_segment(const std::string& segment_id, double timestep);
+    void at_segment(const std::string& segment_id, int timestep);
 
     /**
      * Get the timesteps since a segment was last seen
@@ -49,7 +49,7 @@ namespace types {
      * @param  current_ts the current timestep
      * @return            the number of timesteps since being on that segment or -1 if never seen
      */
-    double timesteps_since_seen(const std::string& segment_id, double current_ts) const;
+    int timesteps_since_seen(const std::string& segment_id, int current_ts) const;
   };
 }
